Outsource loadMessages function in own file
