+++ site/src/app/block-preview/[domain]/[language]/layout.tsx
@@ -1,5 +1,5 @@
-import { loadMessages } from "@src/app/[domain]/[language]/layout";
 import { IntlProvider } from "@src/util/IntlProvider";
+import { loadMessages } from "@src/util/loadMessages";
 import { PropsWithChildren } from "react";
 
 export default async function Page({ children, params: { language } }: PropsWithChildren<{ params: { language: string } }>) {
